Completely rewritten auth-service. Changed name of the service to auth-server
