+++ pom.xml
@@ -13,7 +13,7 @@
 
     <modules>
         <module>socialnetwork-config-server</module>
-        <module>socialnetwork-auth-service</module>
+        <module>socialnetwork-auth-server</module>
         <module>socialnetwork-profile-service</module>
         <module>socialnetwork-api-gateway</module>
         <module>socialnetwork-service-discovery</module>
@@ -25,7 +25,7 @@
         <java.version>14</java.version>
         <spring.cloud.netflix.eureka.server.version>3.1.2</spring.cloud.netflix.eureka.server.version>
         <spring.cloud.openfeign.version>2.2.5.RELEASE</spring.cloud.openfeign.version>
-        <postgresql.version>42.3.3</postgresql.version>
+        <postgresql.version>42.5.4</postgresql.version>
         <google.gson.version>2.8.9</google.gson.version>
         <maven.compiler.source>14</maven.compiler.source>
         <maven.compiler.target>14</maven.compiler.target>
