+++ socialnetwork-auth-server/src/main/java/local/socialnetwork/authserver/SpringBootRunAuthServer.java
@@ -1,0 +1,17 @@
+package local.socialnetwork.authserver;
+
+import org.springframework.boot.SpringApplication;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@EnableFeignClients
+@SpringBootApplication
+public class SpringBootRunAuthServer {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SpringBootRunAuthServer.class);
+    }
+}
