+++ socialnetwork-auth-server/src/main/java/local/socialnetwork/authserver/config/jwt/JwtConfigurer.java
@@ -1,0 +1,29 @@
+package local.socialnetwork.authserver.config.jwt;
+
+import lombok.AccessLevel;
+
+import lombok.experimental.FieldDefaults;
+
+import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
+
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+
+import org.springframework.security.web.DefaultSecurityFilterChain;
+
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class JwtConfigurer extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
+
+    final JwtTokenProvider jwtTokenProvider;
+
+    public JwtConfigurer(JwtTokenProvider jwtTokenProvider) {
+        this.jwtTokenProvider = jwtTokenProvider;
+    }
+
+    @Override
+    public void configure(HttpSecurity http) {
+        JwtTokenFilter customFilter = new JwtTokenFilter(jwtTokenProvider);
+        http.addFilterBefore(customFilter, UsernamePasswordAuthenticationFilter.class);
+    }
+}
