+++ socialnetwork-auth-server/src/main/java/local/socialnetwork/authserver/config/jwt/JwtTokenFilter.java
@@ -1,0 +1,39 @@
+package local.socialnetwork.authserver.config.jwt;
+
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+
+import lombok.AccessLevel;
+
+import lombok.experimental.FieldDefaults;
+
+import org.springframework.security.core.Authentication;
+
+import org.springframework.security.core.context.SecurityContextHolder;
+
+import org.springframework.web.filter.GenericFilterBean;
+
+import java.io.IOException;
+
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class JwtTokenFilter extends GenericFilterBean {
+
+    final JwtTokenProvider jwtTokenProvider;
+
+    JwtTokenFilter(JwtTokenProvider jwtTokenProvider) {
+        this.jwtTokenProvider = jwtTokenProvider;
+    }
+
+    @Override
+    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
+        String token = jwtTokenProvider.resolveToken((HttpServletRequest) servletRequest);
+        if (token != null && jwtTokenProvider.validateToken(token)) {
+            Authentication auth = jwtTokenProvider.getAuthentication(token);
+            SecurityContextHolder.getContext().setAuthentication(auth);
+        }
+        filterChain.doFilter(servletRequest, servletResponse);
+    }
+}
