+++ socialnetwork-auth-server/src/main/java/local/socialnetwork/authserver/config/security/SecurityConfig.java
@@ -1,0 +1,66 @@
+package local.socialnetwork.authserver.config.security;
+
+import local.socialnetwork.authserver.config.jwt.JwtConfigurer;
+import local.socialnetwork.authserver.config.jwt.JwtTokenProvider;
+
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+
+import lombok.experimental.FieldDefaults;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import org.springframework.http.HttpMethod;
+
+import org.springframework.security.authentication.AuthenticationManager;
+
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+
+import org.springframework.security.config.http.SessionCreationPolicy;
+
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+@EnableWebSecurity
+@FieldDefaults(level = AccessLevel.PRIVATE)
+@RequiredArgsConstructor
+public class SecurityConfig {
+
+    final JwtTokenProvider jwtTokenProvider;
+
+    @Bean
+    public SecurityFilterChain configure(HttpSecurity http) throws Exception {
+        http.csrf().disable()
+            .cors().disable()
+            .httpBasic().disable()
+            .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+            .and()
+            .securityMatcher("/auth/**")
+            .authorizeHttpRequests(authorizeRequests ->
+                    authorizeRequests.requestMatchers(HttpMethod.POST, "/registration").permitAll()
+                            .requestMatchers(HttpMethod.POST, "/signin").permitAll()
+                            .anyRequest().authenticated())
+            .apply(new JwtConfigurer(jwtTokenProvider));
+        return http.build();
+    }
+
+    @Bean
+    public AuthenticationManager authenticationManager(HttpSecurity http) throws Exception {
+        AuthenticationManagerBuilder authenticationManagerBuilder = http.getSharedObject(AuthenticationManagerBuilder.class);
+        return authenticationManagerBuilder.build();
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+}
