+++ socialnetwork-auth-server/src/main/java/local/socialnetwork/authserver/dto/AuthenticationUserDTO.java
@@ -1,0 +1,23 @@
+package local.socialnetwork.authserver.dto;
+
+public class AuthenticationUserDTO {
+
+    private String username;
+    private String password;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
