+++ socialnetwork-auth-server/src/main/java/local/socialnetwork/authserver/dto/RegistrationDTO.java
@@ -1,0 +1,17 @@
+package local.socialnetwork.authserver.dto;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.experimental.FieldDefaults;
+
+import java.util.UUID;
+
+@Getter
+@Setter
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class RegistrationDTO {
+    UUID id;
+    String username;
+    String password;
+}
