+++ socialnetwork-auth-server/src/main/java/local/socialnetwork/authserver/exception/UserNotFoundException.java
@@ -1,0 +1,13 @@
+package local.socialnetwork.authserver.exception;
+
+import java.io.Serial;
+
+public class UserNotFoundException extends RuntimeException {
+
+    @Serial
+    private static final long serialVersionUID = 6296010907297189435L;
+
+    public UserNotFoundException(String message) {
+        super(message);
+    }
+}
