+++ socialnetwork-auth-server/src/main/java/local/socialnetwork/authserver/model/entity/AbstractBaseModel.java
@@ -1,0 +1,26 @@
+package local.socialnetwork.authserver.model.entity;
+
+import jakarta.persistence.Id;
+import jakarta.persistence.MappedSuperclass;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+
+import lombok.experimental.FieldDefaults;
+
+import java.util.UUID;
+
+@Getter
+@Setter
+@FieldDefaults(level = AccessLevel.PRIVATE)
+@MappedSuperclass
+public class AbstractBaseModel {
+
+    @Id
+    UUID id;
+
+    public AbstractBaseModel() {
+        this.id = UUID.randomUUID();
+    }
+}
