+++ socialnetwork-auth-server/src/main/java/local/socialnetwork/authserver/model/entity/User.java
@@ -1,0 +1,36 @@
+package local.socialnetwork.authserver.model.entity;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+
+import local.socialnetwork.authserver.model.entity.AbstractBaseModel;
+import local.socialnetwork.authserver.model.entity.Role;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+
+import lombok.experimental.FieldDefaults;
+
+import java.util.List;
+
+@Table(name = "sn_users")
+@Entity
+@Setter
+@Getter
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class User extends AbstractBaseModel {
+
+    @Column(name = "username", nullable = false)
+    String username;
+
+    @Column(name = "password", nullable = false)
+    String password;
+
+    @OneToMany(mappedBy = "user", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+    List<Role> roles;
+
+}
