+++ socialnetwork-auth-server/src/main/java/local/socialnetwork/authserver/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package local.socialnetwork.authserver.repository;
+
+import local.socialnetwork.authserver.model.entity.User;
+
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface UserRepository extends CrudRepository<User, UUID> {
+    Optional<User> findByUsername(String username);
+}
