+++ socialnetwork-auth-server/src/main/resources/application.properties
@@ -1,0 +1,27 @@
+# Application name
+spring.application.name=auth-server
+
+# Application port
+server.port=8080
+
+# Profile
+spring.profiles.active=local
+
+# Spring Cloud: Config
+spring.config.import=optional:configserver:http://localhost:8888
+
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.generate-ddl=true
+spring.jpa.show-sql=true
+
+# Spring: PostgreSQL Configuration
+spring.datasource.url=jdbc:postgresql://localhost:5432/sn_user_service_db
+spring.datasource.username=sn_user_service_user
+spring.datasource.password=sn_user_service_password
+
+# Spring Cloud: Eureka Configuration
+eureka.client.serviceUrl.defaultZone=http://localhost:8761/eureka
+
+# Jwt Configuration
+security.jwt.token.secret-key=c1NxnyGc34E87RnTkERjlwmjkD7Yt/r0XrfwnqcS
+security.jwt.token.expire-length=3600000
