+++ socialnetwork-auth-server/wait-for-it.sh
