--- socialnetwork-auth-service/src/main/java/local/socialnetwork/authservice/SpringBootRunAuthService.java
@@ -1,19 +1,0 @@
-package local.socialnetwork.authservice;
-
-import org.springframework.boot.SpringApplication;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@EnableFeignClients
-@EnableEurekaClient
-@SpringBootApplication
-public class SpringBootRunAuthService {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SpringBootRunAuthService.class);
-    }
-}
