--- socialnetwork-auth-service/src/main/java/local/socialnetwork/authservice/config/jwt/JwtConfigurer.java
@@ -1,27 +1,0 @@
-package local.socialnetwork.authservice.config.jwt;
-
-import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
-
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-
-import org.springframework.security.web.DefaultSecurityFilterChain;
-
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-public class JwtConfigurer extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
-
-    private final JwtTokenProvider jwtTokenProvider;
-
-    public JwtConfigurer(JwtTokenProvider jwtTokenProvider) {
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
-
-    @Override
-    public void configure(HttpSecurity http) {
-
-        JwtTokenFilter customFilter = new JwtTokenFilter(jwtTokenProvider);
-
-        http.addFilterBefore(customFilter, UsernamePasswordAuthenticationFilter.class);
-
-    }
-}
