--- socialnetwork-auth-service/src/main/java/local/socialnetwork/authservice/config/jwt/JwtTokenFilter.java
@@ -1,41 +1,0 @@
-package local.socialnetwork.authservice.config.jwt;
-
-import org.springframework.security.core.Authentication;
-
-import org.springframework.security.core.context.SecurityContextHolder;
-
-import org.springframework.web.filter.GenericFilterBean;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-
-import javax.servlet.http.HttpServletRequest;
-
-import java.io.IOException;
-
-public class JwtTokenFilter extends GenericFilterBean {
-
-    private final JwtTokenProvider jwtTokenProvider;
-
-    JwtTokenFilter(JwtTokenProvider jwtTokenProvider) {
-        this.jwtTokenProvider = jwtTokenProvider;
-    }
-    @Override
-    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
-
-        String token = jwtTokenProvider.resolveToken((HttpServletRequest) servletRequest);
-
-        if (token != null && jwtTokenProvider.validateToken(token)) {
-
-            Authentication auth = jwtTokenProvider.getAuthentication(token);
-
-            SecurityContextHolder.getContext().setAuthentication(auth);
-
-        }
-
-        filterChain.doFilter(servletRequest, servletResponse);
-
-    }
-}
