--- socialnetwork-auth-service/src/main/java/local/socialnetwork/authservice/config/security/SecurityInitializer.java
@@ -1,7 +1,0 @@
-package local.socialnetwork.authservice.config.security;
-
-import org.springframework.security.web.context.AbstractSecurityWebApplicationInitializer;
-
-public class SecurityInitializer extends AbstractSecurityWebApplicationInitializer {
-
-}
