--- socialnetwork-auth-service/src/main/java/local/socialnetwork/authservice/constants/ApplicationConstants.java
@@ -1,19 +1,0 @@
-package local.socialnetwork.authservice.constants;
-
-public final class ApplicationConstants {
-
-    public static final String HOST = "http://socialnetwork-api-gateway:8762";
-
-    private ApplicationConstants() {
-
-    }
-
-    public static final class Services {
-
-        public static final String USER_SERVICE_URL = "user-service";
-
-        private Services() {
-
-        }
-    }
-}
