--- socialnetwork-auth-service/src/main/java/local/socialnetwork/authservice/dto/AuthenticationUserDto.java
@@ -1,23 +1,0 @@
-package local.socialnetwork.authservice.dto;
-
-public class AuthenticationUserDto {
-
-    private String username;
-    private String password;
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
