--- socialnetwork-auth-service/src/main/java/local/socialnetwork/authservice/exception/AuthenticationUserException.java
@@ -1,12 +1,0 @@
-package local.socialnetwork.authservice.exception;
-
-import org.springframework.security.core.AuthenticationException;
-
-public class AuthenticationUserException extends AuthenticationException {
-
-    private static final long serialVersionUID = -1243226497368681892L;
-
-    public AuthenticationUserException(String msg) {
-        super(msg);
-    }
-}
