--- socialnetwork-auth-service/src/main/java/local/socialnetwork/authservice/exception/InvalidJwtAuthenticationException.java
@@ -1,12 +1,0 @@
-package local.socialnetwork.authservice.exception;
-
-import org.springframework.security.core.AuthenticationException;
-
-public class InvalidJwtAuthenticationException extends AuthenticationException {
-
-    private static final long serialVersionUID = -1444026248179450950L;
-
-    public InvalidJwtAuthenticationException(String message) {
-        super(message);
-    }
-}
