--- socialnetwork-auth-service/src/main/java/local/socialnetwork/authservice/model/dto/RoleDto.java
@@ -1,25 +1,0 @@
-package local.socialnetwork.authservice.model.dto;
-
-import java.util.UUID;
-
-public class RoleDto {
-
-    private UUID id;
-    private String authority;
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
-    }
-
-    public String getAuthority() {
-        return authority;
-    }
-
-    public void setAuthority(String authority) {
-        this.authority = authority;
-    }
-}
