--- socialnetwork-auth-service/src/main/java/local/socialnetwork/authservice/model/dto/UserDetailsDto.java
@@ -1,59 +1,0 @@
-package local.socialnetwork.authservice.model.dto;
-
-public class UserDetailsDto {
-
-    private String country;
-    private String city;
-    private String address;
-    private String phone;
-    private String birthday;
-    private String familyStatus;
-
-    public String getCountry() {
-        return country;
-    }
-
-    public void setCountry(String country) {
-        this.country = country;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    public String getBirthday() {
-        return birthday;
-    }
-
-    public void setBirthday(String birthday) {
-        this.birthday = birthday;
-    }
-
-    public String getFamilyStatus() {
-        return familyStatus;
-    }
-
-    public void setFamilyStatus(String familyStatus) {
-        this.familyStatus = familyStatus;
-    }
-}
