--- socialnetwork-auth-service/src/main/java/local/socialnetwork/authservice/model/dto/UserDto.java
@@ -1,71 +1,0 @@
-package local.socialnetwork.authservice.model.dto;
-
-import java.util.List;
-import java.util.UUID;
-
-public class UserDto {
-
-    private UUID id;
-    private String firstName;
-    private String lastName;
-    private String username;
-    private String password;
-    private UserDetailsDto userDetails;
-    private List<RoleDto> roles;
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public UserDetailsDto getUserDetails() {
-        return userDetails;
-    }
-
-    public void setUserDetails(UserDetailsDto userDetails) {
-        this.userDetails = userDetails;
-    }
-
-    public List<RoleDto> getRoles() {
-        return roles;
-    }
-
-    public void setRoles(List<RoleDto> roles) {
-        this.roles = roles;
-    }
-}
