--- socialnetwork-auth-service/src/main/resources/application.properties
@@ -1,11 +1,0 @@
-# Application name
-spring.application.name=auth-service
-
-# Application port
-server.port=8080
-
-# Profile
-spring.profiles.active=local
-
-# Spring Cloud: Config
-spring.config.import=configserver:http://socialnetwork-config-server:8888
